nixl/infra: Move metadata files to utils/stream
Metadata streamer files are now in utils/stream similar for the tests.
The header is used by some tests as well so adjust paths.
