+++ src/utils/stream/metadata_stream.cpp
