+++ src/utils/stream/metadata_stream.h
