+++ test/nixl/meson.build
@@ -38,7 +38,7 @@ agent_example = executable('agent_example',
            install: true)
 
 nixl_ucx_app  = executable('nixl_test', 'nixl_test.cpp',
-                           dependencies: [nixl_dep, nixl_infra] + cuda_dependencies,
+                           dependencies: [nixl_dep, nixl_infra, stream_interface] + cuda_dependencies,
                            include_directories: [nixl_inc_dirs, utils_inc_dirs, '../../src/utils/serdes'],
                            link_with: [serdes_lib], install: true)
